Set content-length
