+++ lib/services/core/servicebusserviceclient.js
@@ -15,7 +15,6 @@
 
 // Module dependencies.
 var util = require('util');
-var url = require('url');
 
 var ServiceClient = require('./serviceclient');
 var Wrap = require('../serviceBus/internal/wrap');
@@ -59,19 +58,8 @@ util.inherits(ServiceBusServiceClient, ServiceClient);
 * @param {object}      options     The request options.
 * @param {function(error, requestOptions)}  callback  The callback function.
 */
-ServiceBusServiceClient.prototype._buildRequestOptions = function (webResource, options, callback) {
-  var self = this;
-
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_TYPE]) {
-    webResource.withHeader(HeaderConstants.CONTENT_TYPE, '');
-  }
-
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_LENGTH]) {
-    webResource.withHeader(HeaderConstants.CONTENT_LENGTH, 0);
-  }
-
+ServiceBusServiceClient.prototype._buildRequestOptions = function (webResource, body, options, callback) {
   webResource.withHeader(HeaderConstants.ACCEPT_CHARSET_HEADER, 'UTF-8');
-  webResource.withHeader(HeaderConstants.HOST_HEADER, this.host + ':' + this.port);
 
   // Set API version
   if (webResource.queryString[Constants.ServiceBusConstants.API_VERSION_QUERY_KEY] === undefined) {
@@ -79,9 +67,6 @@ ServiceBusServiceClient.prototype._buildRequestOptions = function (webResource, 
   } else if (webResource.queryString[Constants.ServiceBusConstants.API_VERSION_QUERY_KEY] === null) {
     delete webResource.queryString[Constants.ServiceBusConstants.API_VERSION_QUERY_KEY];
   }
-
-  // Sets the request url in the web resource.
-  this._setRequestUrl(webResource);
 
   // If wrap is used, make sure proxy settings are in sync
   if (this.useProxy &&
@@ -91,47 +76,7 @@ ServiceBusServiceClient.prototype._buildRequestOptions = function (webResource, 
     this.authenticationProvider.wrapTokenManager.wrapService.setProxy(this.proxyUrl, this.proxyPort);
   }
 
-  // Now that the web request is finalized, sign it
-  this.authenticationProvider.signRequest(webResource, function (error) {
-    var requestOptions = null;
-
-    if (!error) {
-      requestOptions = {
-        url: url.format({
-          protocol: self._isHttps() ? 'https:' : 'http:',
-          hostname: self.host,
-          port: self.port,
-          pathname: webResource.path,
-          query: webResource.queryString
-        }),
-        method: webResource.httpVerb,
-        headers: webResource.headers,
-        strictSSL: self.strictSSL
-      };
-
-      self._setRequestOptionsProxy(requestOptions);
-    }
-
-    callback(error, requestOptions);
-  });
-};
-
-/**
-* Retrieves the normalized path to be used in a request.
-* It adds a leading "/" to the path in case
-* it's not there before.
-*
-* @param {string} path The path to be normalized.
-* @return {string} The normalized path.
-*/
-ServiceBusServiceClient.prototype._getPath = function (path) {
-  if (path === null || path === undefined) {
-    path = '/';
-  } else if (path.indexOf('/') !== 0) {
-    path = '/' + path;
-  }
-
-  return path;
+  ServiceBusServiceClient['super_'].prototype._buildRequestOptions.call(this, webResource, body, options, callback);
 };
 
 module.exports = ServiceBusServiceClient;
