+++ lib/services/core/serviceclient.js
@@ -15,11 +15,13 @@
 */
 
 // Module dependencies.
+var https = require('https');
 var request = require('request');
 var url = require('url');
 var util = require('util');
 var xml2js = require('xml2js');
 var events = require('events');
+var tunnel = require('tunnel');
 var _ = require('underscore');
 
 var azureutil = require('../../util/util');
@@ -187,7 +189,7 @@ ServiceClient.prototype.performRequestInputStream = function (webResource, outpu
 */
 ServiceClient.prototype._performRequest = function (webResource, body, options, callback, chunkedCallback) {
   var self = this;
-  self._buildRequestOptions(webResource, options, function (err, requestOptions) {
+  self._buildRequestOptions(webResource, body, options, function (err, requestOptions) {
     if (err) {
       callback({ error: err, response: null }, function (requestOptions, finalCallback) {
         finalCallback(requestOptions);
@@ -322,6 +324,124 @@ ServiceClient.prototype._performRequest = function (webResource, body, options, 
       });
     }
   });
+};
+
+/**
+* Builds the request options to be passed to the http.request method.
+*
+* @param {WebResource} webResource The webresource where to build the options from.
+* @param {object}      options     The request options.
+* @param {function(error, requestOptions)}  callback  The callback function.
+* @return {undefined}
+*/
+ServiceClient.prototype._buildRequestOptions = function (webResource, body, options, callback) {
+  var self = this;
+
+  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_TYPE]) {
+    webResource.withHeader(HeaderConstants.CONTENT_TYPE, '');
+  }
+
+  if (!webResource.headers || webResource.headers[HeaderConstants.CONTENT_LENGTH] === undefined) {
+    if (body && body.outputData) {
+      webResource.withHeader(HeaderConstants.CONTENT_LENGTH, Buffer.byteLength(body.outputData, 'UTF8'));
+    } else {
+      webResource.withHeader(HeaderConstants.CONTENT_LENGTH, 0);
+    }
+  } else if (webResource.headers && webResource.headers[HeaderConstants.CONTENT_LENGTH] === null) {
+    delete webResource.headers[HeaderConstants.CONTENT_LENGTH];
+  }
+
+  webResource.withHeader(HeaderConstants.ACCEPT_CHARSET_HEADER, 'UTF-8');
+  webResource.withHeader(HeaderConstants.HOST_HEADER, self.host + ':' + self.port);
+
+  // Sets the request url in the web resource.
+  this._setRequestUrl(webResource);
+
+  // Now that the web request is finalized, sign it
+  this.authenticationProvider.signRequest(webResource, function (error) {
+    var requestOptions = null;
+
+    if (!error) {
+      var targetUrl = {
+        protocol: self._isHttps() ? 'https' : 'http',
+        hostname: self.host,
+        port: self.port,
+        pathname: webResource.path,
+        query: webResource.queryString
+      };
+
+      if (webResource.authentication && webResource.authentication.user) {
+        targetUrl.auth = util.format('%s:%s', webResource.authentication.user, webResource.authentication.pass);
+      }
+
+      requestOptions = {
+        url: url.format(targetUrl),
+        method: webResource.httpVerb,
+        headers: webResource.headers,
+        strictSSL: self.strictSSL
+      };
+
+      if (webResource.authentication && webResource.authentication.keyvalue && webResource.authentication.certvalue) {
+        requestOptions.key = webResource.authentication.keyvalue;
+        requestOptions.cert = webResource.authentication.certvalue;
+      }
+
+      self._setRequestOptionsProxy(requestOptions);
+      self._setAgent(self, requestOptions, self.protocol.toLowerCase() === Constants.HTTPS);
+
+      if(options) {
+        //set encoding of response data. If set to null, the body is returned as a Buffer
+        requestOptions.encoding = options.responseEncoding;
+        //set client request time out
+        if(options.clientRequestTimeout && options.clientRequestTimeout > 0) {
+          requestOptions.timeout = options.clientRequestTimeout;
+        } else {
+          requestOptions.timeout = Constants.DEFAULT_CLIENT_REQUEST_TIMEOUT;
+        }
+      }
+    }
+
+    callback(error, requestOptions);
+  });
+};
+
+/**
+* Set the Agent to use for the request
+*  Result depends on proxy settings and protocol
+*
+* @param {object}   reqopts     request options for request
+* @param {bool}     isHTTPS     true - use https to proxy. Otherwise use http.
+*/
+ServiceClient.prototype._setAgent = function (self, reqopts, isHTTPS) {
+  if (self.useTunnelProxy && self.proxy) {
+    var agentinfo = {
+      proxy: self.proxy
+    };
+    if (reqopts.key) {
+      agentinfo.key = reqopts.key;
+    }
+    if (reqopts.cert) {
+      agentinfo.cert = reqopts.cert;
+    }
+    if (this.maxSockets) {
+      agentinfo.maxSockets = self.maxSockets;
+    }
+    if (isHTTPS) {
+      if (self.proxyIsHTTPS) {
+        reqopts.agent = tunnel.httpsOverHttps(agentinfo);
+      } else {
+        reqopts.agent = tunnel.httpsOverHttp(agentinfo);
+      }
+    } else {
+      if (self.proxyIsHTTPS) {
+        reqopts.agent = tunnel.httpOverHttps(agentinfo);
+      } else {
+        reqopts.agent = tunnel.httpOverHttp(agentinfo);
+      }
+    }
+  } else if (isHTTPS) {
+    reqopts.agent = new https.Agent(reqopts);
+  }
 };
 
 ServiceClient.prototype._tagRequest = function (requestOptions) {
@@ -636,6 +756,24 @@ ServiceClient.prototype._setRequestUrl = function (webResource) {
     pathname: webResource.path,
     query: webResource.queryString
   });
+};
+
+/**
+* Retrieves the normalized path to be used in a request.
+* It adds a leading "/" to the path in case
+* it's not there before.
+*
+* @param {string} path The path to be normalized.
+* @return {string} The normalized path.
+*/
+ServiceClient.prototype._getPath = function (path) {
+  if (path === null || path === undefined) {
+    path = '/';
+  } else if (path.indexOf('/') !== 0) {
+    path = '/' + path;
+  }
+
+  return path;
 };
 
 /**
