+++ lib/services/core/servicemanagementclient.js
@@ -14,11 +14,10 @@
 */
 
 // Module dependencies.
-var https = require('https');
 var util = require('util');
 var fs = require('fs');
 var url = require('url');
-var tunnel = require('tunnel');
+
 var ServiceClient = require('./serviceclient');
 var Constants = require('../../util/constants');
 var ServiceClientConstants = require('./serviceclientconstants');
@@ -47,7 +46,21 @@ function ServiceManagementClient(settingsOrAuthentication, hostOptions) {
     this._setAuthentication(settingsOrAuthentication);
     this._setServiceHost(hostOptions);
   }
+
   this._setDefaultProxy();
+
+  var self = this;
+
+  this.authenticationProvider = {
+    signRequest: function (webResource, callback) {
+      webResource.authentication = {
+        keyvalue: self.keyvalue,
+        certvalue: self.certvalue
+      };
+
+      callback(null);
+    }
+  };
 }
 
 util.inherits(ServiceManagementClient, ServiceClient);
@@ -158,9 +171,7 @@ ServiceManagementClient.prototype._setServiceHostFromSettings = function(setting
 * @param {function(error, requestOptions)}  callback  The callback function.
 * @return {undefined}
 */
-ServiceManagementClient.prototype._buildRequestOptions = function (webResource, options, callback) {
-  var self = this;
-
+ServiceManagementClient.prototype._buildRequestOptions = function (webResource, body, options, callback) {
   if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_TYPE]) {
     webResource.withHeader(HeaderConstants.CONTENT_TYPE, 'application/xml');
   }
@@ -170,32 +181,8 @@ ServiceManagementClient.prototype._buildRequestOptions = function (webResource, 
   }
 
   webResource.withHeader(HeaderConstants.ACCEPT_CHARSET_HEADER, 'UTF-8');
-  webResource.withHeader(HeaderConstants.STORAGE_VERSION_HEADER, self.apiversion);
-  webResource.withHeader(HeaderConstants.HOST_HEADER, self.host);
 
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_LENGTH]) {
-    webResource.withHeader(HeaderConstants.CONTENT_LENGTH, 0);
-  }
-
-  var requestOptions = {
-    url: url.format({
-      protocol: self._isHttps() ? 'https' : 'http',
-      hostname: self.host,
-      port: self.port,
-      pathname: webResource.path,
-      query: webResource.queryString
-    }),
-    method: webResource.httpVerb,
-    headers: webResource.headers,
-    key: self.keyvalue,
-    cert: self.certvalue,
-    strictSSL: self.strictSSL
-  };
-
-  self._setRequestOptionsProxy(requestOptions);
-
-  self._setAgent(self, requestOptions, self.protocol.toLowerCase() === Constants.HTTPS);
-  callback(null, requestOptions);
+  ServiceManagementClient['super_'].prototype._buildRequestOptions.call(this, webResource, body, options, callback);
 };
 
 /**
@@ -254,45 +241,6 @@ ServiceManagementClient.prototype._setProxy = function (proxy, isHTTPS) {
   } else {
     this.useTunnelProxy = false;
     this.proxy = null;
-  }
-};
-
-/**
-* Set the Agent to use for the request
-*  Result depends on proxy settings and protocol
-*
-* @param {object}   reqopts     request options for request
-* @param {bool}     isHTTPS     true - use https to proxy. Otherwise use http.
-*/
-ServiceManagementClient.prototype._setAgent = function (self, reqopts, isHTTPS) {
-  if (self.useTunnelProxy && self.proxy) {
-    var agentinfo = {
-      proxy: self.proxy
-    };
-    if (reqopts.key) {
-      agentinfo.key = reqopts.key;
-    }
-    if (reqopts.cert) {
-      agentinfo.cert = reqopts.cert;
-    }
-    if (this.maxSockets) {
-      agentinfo.maxSockets = self.maxSockets;
-    }
-    if (isHTTPS) {
-      if (self.proxyIsHTTPS) {
-        reqopts.agent = tunnel.httpsOverHttps(agentinfo);
-      } else {
-        reqopts.agent = tunnel.httpsOverHttp(agentinfo);
-      }
-    } else {
-      if (self.proxyIsHTTPS) {
-        reqopts.agent = tunnel.httpOverHttps(agentinfo);
-      } else {
-        reqopts.agent = tunnel.httpOverHttp(agentinfo);
-      }
-    }
-  } else if (isHTTPS) {
-    reqopts.agent = new https.Agent(reqopts);
   }
 };
