+++ lib/services/core/storageserviceclient.js
@@ -15,7 +15,6 @@
 
 // Module dependencies.
 var _ = require('underscore');
-var url = require('url');
 var util = require('util');
 
 var azureutil = require('../../util/util');
@@ -89,25 +88,11 @@ StorageServiceClient.getStorageSettings = function (storageAccountOrConnectionSt
 * @param {function(error, requestOptions)}  callback  The callback function.
 * @return {undefined}
 */
-StorageServiceClient.prototype._buildRequestOptions = function (webResource, options, callback) {
-  var self = this;
-
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_TYPE]) {
-    webResource.withHeader(HeaderConstants.CONTENT_TYPE, '');
-  }
-
-  if (!webResource.headers || webResource.headers[HeaderConstants.CONTENT_LENGTH] === undefined) {
-    webResource.withHeader(HeaderConstants.CONTENT_LENGTH, 0);
-  } else if (webResource.headers && webResource.headers[HeaderConstants.CONTENT_LENGTH] === null) {
-    delete webResource.headers[HeaderConstants.CONTENT_LENGTH];
-  }
-
+StorageServiceClient.prototype._buildRequestOptions = function (webResource, body, options, callback) {
   webResource.withHeader(HeaderConstants.STORAGE_VERSION_HEADER, this.apiVersion);
   webResource.withHeader(HeaderConstants.DATE_HEADER, new Date().toUTCString());
   webResource.withHeader(HeaderConstants.ACCEPT_HEADER, 'application/atom+xml,application/xml');
   webResource.withHeader(HeaderConstants.ACCEPT_CHARSET_HEADER, 'UTF-8');
-
-  webResource.withHeader(HeaderConstants.HOST_HEADER, this.host + ':' + this.port);
 
   if (options) {
     if (options.timeoutIntervalInMs) {
@@ -131,43 +116,7 @@ StorageServiceClient.prototype._buildRequestOptions = function (webResource, opt
     }
   }
 
-  // Sets the request url in the web resource.
-  this._setRequestUrl(webResource);
-
-  // Now that the web request is finalized, sign it
-  this.authenticationProvider.signRequest(webResource, function (error) {
-    var requestOptions = null;
-
-    if (!error) {
-      requestOptions = {
-        url: url.format({
-          protocol: self._isHttps() ? 'https:' : 'http:',
-          hostname: self.host,
-          port: self.port,
-          pathname: webResource.path,
-          query: webResource.queryString
-        }),
-        method: webResource.httpVerb,
-        headers: webResource.headers,
-        strictSSL: self.strictSSL
-      };
-
-      if(options) {
-        //set encoding of response data. If set to null, the body is returned as a Buffer
-        requestOptions.encoding = options.responseEncoding;
-        //set client request time out
-        if(options.clientRequestTimeout && options.clientRequestTimeout > 0) {
-          requestOptions.timeout = options.clientRequestTimeout;
-        } else {
-          requestOptions.timeout = Constants.DEFAULT_CLIENT_REQUEST_TIMEOUT;
-        }
-      }
-
-      self._setRequestOptionsProxy(requestOptions);
-    }
-
-    callback(error, requestOptions);
-  });
+  StorageServiceClient['super_'].prototype._buildRequestOptions.call(this, webResource, body, options, callback);
 };
 
 /**
