+++ lib/services/scm/scmservice.js
@@ -14,10 +14,11 @@
 */
 
 // Module dependencies.
+var url = require('url');
 var util = require('util');
 var azureutil = require('../../util/util');
 
-var ScmServiceClient = require('../core/scmserviceclient');
+var ServiceClient = require('../core/serviceclient');
 var WebResource = require('../../http/webresource');
 
 var Constants = require('../../util/constants');
@@ -38,10 +39,19 @@ var HeaderConstants = Constants.HeaderConstants;
 * @param {string} [hostOptions.host]                                   The SCM repository endpoint.
 */
 function ScmService(authentication, hostOptions) {
-  ScmService['super_'].call(this, authentication, hostOptions);
+  hostOptions.protocol = Constants.HTTPS;
+
+  ScmService['super_'].call(this, url.format(hostOptions), authentication);
+
+  this.authenticationProvider = {
+    signRequest: function (webResource, callback) {
+      webResource.authentication = authentication;
+      callback(null);
+    }
+  };
 }
 
-util.inherits(ScmService, ScmServiceClient);
+util.inherits(ScmService, ServiceClient);
 
 /**
 * Lists the available deployments.
