+++ lib/services/serviceBus/wrapservice.js
@@ -58,6 +58,14 @@ function WrapService(acsHost, issuer, accessKey) {
   }
 
   WrapService['super_'].call(this, acsHost);
+
+  this.authenticationProvider = {
+    signRequest: function (webResource, callback) {
+      callback(null);
+    }
+  };
+
+  this.strictSSL = false;
 }
 
 util.inherits(WrapService, ServiceClient);
@@ -107,53 +115,6 @@ WrapService.prototype.wrapAccessToken = function (uri, optionsOrCallback, callba
   };
 
   this.performRequest(webResource, acsData, options, processResponseCallback);
-};
-
-WrapService.prototype._buildRequestOptions = function (webResource, options, callback) {
-  var self = this;
-
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_TYPE]) {
-    webResource.withHeader(HeaderConstants.CONTENT_TYPE, '');
-  }
-
-  if (!webResource.headers || !webResource.headers[HeaderConstants.CONTENT_LENGTH]) {
-    webResource.withHeader(HeaderConstants.CONTENT_LENGTH, 0);
-  }
-
-  // Sets the request url in the web resource.
-  this._setRequestUrl(webResource);
-
-  var requestOptions = {
-    url: url.format({
-      protocol: self._isHttps() ? 'https:' : 'http:',
-      hostname: self.host,
-      port: self.port,
-      pathname: webResource.path,
-      query: webResource.queryString
-    }),
-    method: webResource.httpVerb,
-    headers: webResource.headers
-  };
-
-  callback(null, requestOptions);
-};
-
-/**
-* Retrieves the normalized path to be used in a request.
-* It adds a leading "/" to the path in case
-* it's not there before.
-*
-* @param {string} path The path to be normalized.
-* @return {string} The normalized path.
-*/
-WrapService.prototype._getPath = function (path) {
-  if (path === null || path === undefined) {
-    path = '/';
-  } else if (path.indexOf('/') !== 0) {
-    path = '/' + path;
-  }
-
-  return path;
 };
 
 module.exports = WrapService;
