+++ lib/services/sqlAzure/sqlserveracs.js
@@ -89,7 +89,7 @@ SqlServerAcs.prototype.signRequest = function (webResourceToSign, callback) {
   this.performRequest(webResource, null, null, processResponseCallback);
 };
 
-SqlServerAcs.prototype._buildRequestOptions = function (webResource, options, callback) {
+SqlServerAcs.prototype._buildRequestOptions = function (webResource, body, options, callback) {
   var self = this;
 
   // Sets the request url in the web resource.
